+++ backend/footprint/src/main/java/foot/footprint/domain/member/application/auth/AuthService.java
@@ -1,0 +1,8 @@
+package foot.footprint.domain.member.application.auth;
+
+import foot.footprint.domain.member.dto.authRequest.AuthRequest;
+
+public interface AuthService {
+
+    String process(AuthRequest authRequest);
+}
