+++ backend/footprint/src/test/java/foot/footprint/service/member/MemberServiceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import foot.footprint.domain.member.application.MemberServiceImpl;
+import foot.footprint.domain.member.application.member.MemberServiceImpl;
 import foot.footprint.domain.member.dao.MemberRepository;
 import foot.footprint.domain.member.domain.AuthProvider;
 import foot.footprint.domain.member.domain.Member;
